[admin] edito admin contra de mas de 5

+++ EstamperiaSanchezLaravel/resources/views/product/index.blade.php
@@ -32,8 +32,10 @@
 					<img src="{{ $product->image }}" alt="calse remera {{ $product->id }}" class="imgreme">
 					<h5 class="nombreRemera">{{ $product->name }}</h5>
 				</a>
-				<a href="/product/edit/{{$product->id}}"> Editar </a>
-				<a href="/product/delete/{{$product->id}}"> Eliminar</a>
+				@if(\Auth::user()->admin == '1')
+					<a href="/product/edit/{{$product->id}}"> Editar </a>
+					<a href="/product/delete/{{$product->id}}"> Eliminar</a>
+				@endif
 			</div>
 			{{--TERMINA CADA PRODUCTO--}}
 		@empty
